+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to the "vscode-abap-remote-fs" extension will be documented 
 
 Format based on [Keep a Changelog](http://keepachangelog.com/)
 
+## 4.1.0 [2021-11-05]
+
+### Added
+
+- request logging
+
+### Fixed
+
+- stateful session handling
+- cookie handling
+
 ## 4.0.0 [2021-11-02]
 
 ### Changed